Extract logic to fetch entity suggestions into custom hook
This makes the component code less complex and allows us to reuse logic in the future.
